Makefile: Move 'all' back to top so its default
I repositioned the 'deps' target in a previous commit,
that added an ifdef in case REBAR was already defined:
71dd390e8222c0dbbcc376564910ef03210aa46d

However, I didn't realize that Makefiles depend on having
their default targets always at the top (it doesn't just always
magically default to 'all').

Therefore, this restores the intended order of the targets.