Add include_apps option to allow including additional dependencies to rebar3 coverage reports
This change introduces the `include_apps` option to the rebar3 covertool pluigin, to allow users to specify a list of dependent apps to also generate cobertura coverage reports for. This can be very helpful if you have some libraries that cannot be effectively tested on their own, but are tested via ct in a higher level test.

As the README notes, you'll only get coverage data if you include the beams from the deps in your initital coverage report. However, this allows you to bring the data already included (separately) in the `.coverdata` file into a Cobertura report that is keyed by application name.

Lastly, this fixes a bug introduced by idubrov/covertool#35, where the ebin directory for applications weren't included in the rebar3 coverage report, causing empty reports to be generated.